+++ maintain/release/hooks.py
@@ -1,5 +1,6 @@
 import subprocess
 import logging
+import os
 
 from maintain.release.base import Releaser
 
@@ -25,22 +26,22 @@ class HookReleaser(Releaser):
         self.post_release_commands = config.get('publish', {}).get('post', [])
 
     def pre_bump(self, new_version):
-        self.execute_hooks('pre bump', self.pre_bump_commands, new_version)
+        self.execute_hooks('pre_bump', self.pre_bump_commands, new_version)
 
     def bump(self, new_version):
         self.execute_hooks('bump', self.bump_commands, new_version)
 
     def post_bump(self, new_version):
-        self.execute_hooks('post bump', self.post_bump_commands, new_version)
+        self.execute_hooks('post_bump', self.post_bump_commands, new_version)
 
     def pre_release(self, new_version):
-        self.execute_hooks('pre release', self.pre_release_commands, new_version)
+        self.execute_hooks('pre_release', self.pre_release_commands, new_version)
 
     def release(self, new_version):
         self.execute_hooks('release', self.release_commands, new_version)
 
     def post_release(self, new_version):
-        self.execute_hooks('post release', self.post_release_commands, new_version)
+        self.execute_hooks('post_release', self.post_release_commands, new_version)
 
     def execute_hooks(self, phase, commands, version):
         if len(commands) > 0:
@@ -49,3 +50,8 @@ class HookReleaser(Releaser):
             for hook in commands:
                 logger.info('$ {}'.format(hook))
                 subprocess.check_output(hook, shell=True, env={'VERSION': version})
+
+        hook_file = './.maintain/hooks/{}'.format(phase)
+
+        if os.path.exists(hook_file) and os.access(hook_file, os.X_OK):
+            subprocess.check_output(hook_file, shell=True, env={'VERSION': version})